refactor: try, catch문 적용 및 경로 수정
